split and move API acceptance tests
