+++ flocker/acceptance/obsolete/__init__.py
@@ -1,0 +1,6 @@
+# Copyright Hybrid Logic Ltd.  See LICENSE file for details.
+
+"""
+Acceptance tests for ``flocker`` components that are either deprecated
+or planned for deprection in the near future.
+"""
