+++ flocker/acceptance/obsolete/test_cli.py
